// client/src/constants/theme.ts
export enum COLOR {
    turquoise = '#2a9d8f',
    yellow = '#e9c46a',
    orange = '#f4a261',
    red = '#ef6f51',
    redDark = '#a74d38',
    whiteLight = '#fff',
    white = '#fafafa',
    whiteDark = '#ededed',
    grey = '#747474',
    blackLight = '#424242',
    black = '#303030',
    blueLight = '#4b67f7',
    blue = '#4443d8',
}
export enum MODE {
    LIGHT = 'light',
    DARK = 'dark',
}